requirements/base.txt
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
redis==5.0.6
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg==3.2.1
psycopg-binary==3.2.1